requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
frozenlist==1.4.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
multidict==6.1.0
packaging==24.1
passlib==1.7.4
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.2
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
yarl==1.14.0
